Fixed test failures by removing tests.
