+++ test/Microsoft.ML.Tests/Scenarios/Api/ApiScenariosTests.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.ML.Runtime.Api;
+using Microsoft.ML.Runtime.Data;
 using Microsoft.ML.TestFramework;
 using Xunit.Abstractions;
 
@@ -49,6 +50,55 @@ namespace Microsoft.ML.Tests.Scenarios.Api
             public bool Sentiment;
 
             public float Score;
+        }
+
+        private static TextTransform.Arguments MakeSentimentTextTransformArgs(bool normalize = true)
+        {
+            return new TextTransform.Arguments()
+            {
+                Column = new TextTransform.Column
+                {
+                    Name = "Features",
+                    Source = new[] { "SentimentText" }
+                },
+                OutputTokens = true,
+                KeepPunctuations = false,
+                StopWordsRemover = new Runtime.TextAnalytics.PredefinedStopWordsRemoverFactory(),
+                VectorNormalizer = normalize ? TextTransform.TextNormKind.L2 : TextTransform.TextNormKind.None,
+                CharFeatureExtractor = new NgramExtractorTransform.NgramExtractorArguments() { NgramLength = 3, AllLengths = false },
+                WordFeatureExtractor = new NgramExtractorTransform.NgramExtractorArguments() { NgramLength = 2, AllLengths = true },
+            };
+        }
+
+        private static TextLoader.Arguments MakeIrisTextLoaderArgs()
+        {
+            return new TextLoader.Arguments()
+            {
+                Separator = "comma",
+                HasHeader = true,
+                Column = new[]
+                {
+                    new TextLoader.Column("SepalLength", DataKind.R4, 0),
+                    new TextLoader.Column("SepalWidth", DataKind.R4, 1),
+                    new TextLoader.Column("PetalLength", DataKind.R4, 2),
+                    new TextLoader.Column("PetalWidth",DataKind.R4, 3),
+                    new TextLoader.Column("Label", DataKind.Text, 4)
+                }
+            };
+        }
+
+        private static TextLoader.Arguments MakeSentimentTextLoaderArgs()
+        {
+            return new TextLoader.Arguments()
+            {
+                Separator = "tab",
+                HasHeader = true,
+                Column = new[]
+                {
+                    new TextLoader.Column("Label", DataKind.BL, 0),
+                    new TextLoader.Column("SentimentText", DataKind.Text, 1)
+                }
+            };
         }
     }
 }
