--- test/Microsoft.ML.Tests/Scenarios/Api/DecomposableTrainAndPredict.cs
@@ -1,61 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.ML.Runtime.Api;
-using Microsoft.ML.Runtime.Data;
-using Microsoft.ML.Runtime.RunTests;
-using Microsoft.ML.TestFramework;
-using Microsoft.ML.Trainers;
-using System.Linq;
-using Xunit;
-
-namespace Microsoft.ML.Tests.Scenarios.Api
-{
-
-    public partial class ApiScenariosTests
-    {
-        /// <summary>
-        /// Decomposable train and predict: Train on Iris multiclass problem, which will require
-        /// a transform on labels. Be able to reconstitute the pipeline for a prediction only task,
-        /// which will essentially "drop" the transform over labels, while retaining the property
-        /// that the predicted label for this has a key-type, the probability outputs for the classes
-        /// have the class labels as slot names, etc. This should be do-able without ugly compromises like,
-        /// say, injecting a dummy label.
-        /// </summary>
-        [Fact]
-        void DecomposableTrainAndPredict()
-        {
-            using (var env = new LocalEnvironment()
-                .AddStandardComponents()) // ScoreUtils.GetScorer requires scorers to be registered in the ComponentCatalog
-            {
-                var loader = TextLoader.ReadFile(env, MakeIrisTextLoaderArgs(), new MultiFileSource(GetDataPath(TestDatasets.irisData.trainFilename)));
-                var term = TermTransform.Create(env, loader, "Label");
-                var concat = new ConcatTransform(env, "Features", "SepalLength", "SepalWidth", "PetalLength", "PetalWidth").Transform(term);
-                var trainer = new SdcaMultiClassTrainer(env, "Features", "Label", advancedSettings: (s) => { s.MaxIterations = 100; s.Shuffle = true; s.NumThreads = 1; });
-
-                IDataView trainData = trainer.Info.WantCaching ? (IDataView)new CacheDataView(env, concat, prefetch: null) : concat;
-                var trainRoles = new RoleMappedData(trainData, label: "Label", feature: "Features");
-
-                // Auto-normalization.
-                NormalizeTransform.CreateIfNeeded(env, ref trainRoles, trainer);
-                var predictor = trainer.Train(new Runtime.TrainContext(trainRoles));
-
-                var scoreRoles = new RoleMappedData(concat, label: "Label", feature: "Features");
-                IDataScorerTransform scorer = ScoreUtils.GetScorer(predictor, scoreRoles, env, trainRoles.Schema);
-
-                // Cut out term transform from pipeline.
-                var newScorer = ApplyTransformUtils.ApplyAllTransformsToData(env, scorer, loader, term);
-                var keyToValue = new KeyToValueTransform(env, "PredictedLabel").Transform(newScorer);
-                var model = env.CreatePredictionEngine<IrisDataNoLabel, IrisPrediction>(keyToValue);
-
-                var testData = loader.AsEnumerable<IrisDataNoLabel>(env, false);
-                foreach (var input in testData.Take(20))
-                {
-                    var prediction = model.Predict(input);
-                    Assert.True(prediction.PredictedLabel == "Iris-setosa");
-                }
-            }
-        }
-    }
-}
