--- test/Microsoft.ML.Tests/Scenarios/Api/Evaluation.cs
@@ -1,57 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.ML.Runtime.Api;
-using Microsoft.ML.Runtime.Data;
-using Microsoft.ML.Runtime.RunTests;
-using Microsoft.ML.Trainers;
-using Xunit;
-
-namespace Microsoft.ML.Tests.Scenarios.Api
-{
-    public partial class ApiScenariosTests
-    {
-        /// <summary>
-        /// Evaluation: Similar to the simple train scenario, except instead of having some 
-        /// predictive structure, be able to score another "test" data file, run the result 
-        /// through an evaluator and get metrics like AUC, accuracy, PR curves, and whatnot. 
-        /// Getting metrics out of this shoudl be as straightforward and unannoying as possible.
-        /// </summary>
-        [Fact]
-        public void Evaluation()
-        {
-            using (var env = new LocalEnvironment(seed: 1, conc: 1))
-            {
-                // Pipeline
-                var loader = TextLoader.ReadFile(env, MakeSentimentTextLoaderArgs(), new MultiFileSource(GetDataPath(TestDatasets.Sentiment.trainFilename)));
-
-                var trans = TextTransform.Create(env, MakeSentimentTextTransformArgs(), loader);
-
-                // Train
-                var trainer = new LinearClassificationTrainer(env, new LinearClassificationTrainer.Arguments
-                {
-                    NumThreads = 1
-                });
-
-                var cached = new CacheDataView(env, trans, prefetch: null);
-                var trainRoles = new RoleMappedData(cached, label: "Label", feature: "Features");
-                var predictor = trainer.Train(new Runtime.TrainContext(trainRoles));
-                var scoreRoles = new RoleMappedData(trans, label: "Label", feature: "Features");
-                IDataScorerTransform scorer = ScoreUtils.GetScorer(predictor, scoreRoles, env, trainRoles.Schema);
-
-                // Create prediction engine and test predictions.
-                var model = env.CreatePredictionEngine<SentimentData, SentimentPrediction>(scorer);
-
-                // Take a couple examples out of the test data and run predictions on top.
-                var testLoader = TextLoader.ReadFile(env, MakeSentimentTextLoaderArgs(), new MultiFileSource(GetDataPath(TestDatasets.irisData.testFilename)));
-                var testData = testLoader.AsEnumerable<SentimentData>(env, false);
-
-                var dataEval = new RoleMappedData(scorer, label: "Label", feature: "Features", opt: true);
-
-                var evaluator = new BinaryClassifierMamlEvaluator(env, new BinaryClassifierMamlEvaluator.Arguments() { });
-                var metricsDict = evaluator.Evaluate(dataEval);
-            }
-        }
-    }
-}
