--- test/Microsoft.ML.Tests/Scenarios/Api/SimpleTrainAndPredict.cs
@@ -1,112 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.ML.Runtime.Api;
-using Microsoft.ML.Runtime.Data;
-using Microsoft.ML.Runtime.RunTests;
-using Microsoft.ML.Trainers;
-using System.Linq;
-using Xunit;
-
-namespace Microsoft.ML.Tests.Scenarios.Api
-{
-    public partial class ApiScenariosTests
-    {
-        /// <summary>
-        /// Start with a dataset in a text file. Run text featurization on text values. 
-        /// Train a linear model over that. (I am thinking sentiment classification.) 
-        /// Out of the result, produce some structure over which you can get predictions programmatically 
-        /// (for example, the prediction does not happen over a file as it did during training).
-        /// </summary>
-        [Fact]
-        public void SimpleTrainAndPredict()
-        {
-            var dataset = TestDatasets.Sentiment;
-
-            using (var env = new LocalEnvironment(seed: 1, conc: 1))
-            {
-                // Pipeline
-                var loader = TextLoader.ReadFile(env, MakeSentimentTextLoaderArgs(), new MultiFileSource(GetDataPath(dataset.trainFilename)));
-
-                var trans = TextTransform.Create(env, MakeSentimentTextTransformArgs(), loader);
-
-                // Train
-                var trainer = new LinearClassificationTrainer(env, new LinearClassificationTrainer.Arguments
-                {
-                    NumThreads = 1
-                });
-
-                var cached = new CacheDataView(env, trans, prefetch: null);
-                var trainRoles = new RoleMappedData(cached, label: "Label", feature: "Features");
-                var predictor = trainer.Train(new Runtime.TrainContext(trainRoles));
-
-                var scoreRoles = new RoleMappedData(trans, label: "Label", feature: "Features");
-                IDataScorerTransform scorer = ScoreUtils.GetScorer(predictor, scoreRoles, env, trainRoles.Schema);
-
-                // Create prediction engine and test predictions.
-                var model = env.CreatePredictionEngine<SentimentData, SentimentPrediction>(scorer);
-
-                // Take a couple examples out of the test data and run predictions on top.
-                var testLoader = TextLoader.ReadFile(env, MakeSentimentTextLoaderArgs(), new MultiFileSource(GetDataPath(dataset.testFilename)));
-                var testData = testLoader.AsEnumerable<SentimentData>(env, false);
-                foreach (var input in testData.Take(5))
-                {
-                    var prediction = model.Predict(input);
-                    // Verify that predictions match and scores are separated from zero.
-                    Assert.Equal(input.Sentiment, prediction.Sentiment);
-                    Assert.True(input.Sentiment && prediction.Score > 1 || !input.Sentiment && prediction.Score < -1);
-                }
-            }
-        }
-
-        private static TextTransform.Arguments MakeSentimentTextTransformArgs(bool normalize = true)
-        {
-            return new TextTransform.Arguments()
-            {
-                Column = new TextTransform.Column
-                {
-                    Name = "Features",
-                    Source = new[] { "SentimentText" }
-                },
-                OutputTokens = true,
-                KeepPunctuations=false,
-                StopWordsRemover = new Runtime.TextAnalytics.PredefinedStopWordsRemoverFactory(),
-                VectorNormalizer = normalize ? TextTransform.TextNormKind.L2 : TextTransform.TextNormKind.None,
-                CharFeatureExtractor = new NgramExtractorTransform.NgramExtractorArguments() { NgramLength = 3, AllLengths = false },
-                WordFeatureExtractor = new NgramExtractorTransform.NgramExtractorArguments() { NgramLength = 2, AllLengths = true },
-            };
-        }
-
-        private static TextLoader.Arguments MakeIrisTextLoaderArgs()
-        {
-            return new TextLoader.Arguments()
-            {
-                Separator = "comma",
-                HasHeader = true,
-                Column = new[]
-                {
-                    new TextLoader.Column("SepalLength", DataKind.R4, 0),
-                    new TextLoader.Column("SepalWidth", DataKind.R4, 1),
-                    new TextLoader.Column("PetalLength", DataKind.R4, 2),
-                    new TextLoader.Column("PetalWidth",DataKind.R4, 3),
-                    new TextLoader.Column("Label", DataKind.Text, 4)
-                }
-            };
-        }
-
-        private static TextLoader.Arguments MakeSentimentTextLoaderArgs()
-        {
-            return new TextLoader.Arguments()
-            {
-                Separator = "tab",
-                HasHeader = true,
-                Column = new[]
-                {
-                    new TextLoader.Column("Label", DataKind.BL, 0),
-                    new TextLoader.Column("SentimentText", DataKind.Text, 1)
-                }
-            };
-        }
-    }
-}
