--- test/Microsoft.ML.Tests/Scenarios/Api/TrainSaveModelAndPredict.cs
@@ -1,68 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.ML.Runtime.Api;
-using Microsoft.ML.Runtime.Data;
-using Microsoft.ML.Runtime.RunTests;
-using Microsoft.ML.Trainers;
-using System.Linq;
-using Xunit;
-
-namespace Microsoft.ML.Tests.Scenarios.Api
-{
-    public partial class ApiScenariosTests
-    {
-        /// <summary>
-        /// Train, save/load model, predict: 
-        /// Serve the scenario where training and prediction happen in different processes (or even different machines). 
-        /// The actual test will not run in different processes, but will simulate the idea that the 
-        /// "communication pipe" is just a serialized model of some form.
-        /// </summary>
-        [Fact]
-        public void TrainSaveModelAndPredict()
-        {
-            using (var env = new LocalEnvironment(seed: 1, conc: 1))
-            {
-                // Pipeline
-                var loader = TextLoader.ReadFile(env, MakeSentimentTextLoaderArgs(), new MultiFileSource(GetDataPath(TestDatasets.Sentiment.trainFilename)));
-
-                var trans = TextTransform.Create(env, MakeSentimentTextTransformArgs(), loader);
-
-                // Train
-                var trainer = new LinearClassificationTrainer(env, new LinearClassificationTrainer.Arguments
-                {
-                    NumThreads = 1
-                });
-
-                var cached = new CacheDataView(env, trans, prefetch: null);
-                var trainRoles = new RoleMappedData(cached, label: "Label", feature: "Features");
-                var predictor = trainer.Train(new Runtime.TrainContext(trainRoles));
-
-                PredictionEngine<SentimentData, SentimentPrediction> model;
-                using (var file = env.CreateTempFile())
-                {
-                    // Save model. 
-                    var scoreRoles = new RoleMappedData(trans, label: "Label", feature: "Features");
-                    using (var ch = env.Start("saving"))
-                        TrainUtils.SaveModel(env, ch, file, predictor, scoreRoles);
-
-                    // Load model.
-                    using (var fs = file.OpenReadStream())
-                        model = env.CreatePredictionEngine<SentimentData, SentimentPrediction>(fs);
-                }
-
-                // Take a couple examples out of the test data and run predictions on top.
-                var testLoader = TextLoader.ReadFile(env, MakeSentimentTextLoaderArgs(), new MultiFileSource(GetDataPath(TestDatasets.Sentiment.testFilename)));
-                var testData = testLoader.AsEnumerable<SentimentData>(env, false);
-                foreach (var input in testData.Take(5))
-                {
-                    var prediction = model.Predict(input);
-                    // Verify that predictions match and scores are separated from zero.
-                    Assert.Equal(input.Sentiment, prediction.Sentiment);
-                    Assert.True(input.Sentiment && prediction.Score > 1 || !input.Sentiment && prediction.Score < -1);
-                }
-            }
-        }
-    }
-}
